Fix storage test 'TestPebbleWritesSameSSTs' on big endian
The test 'TestPebbleWritesSameSSTs' within the storage package was failing
on s390x architecture. It was determined that snappy compression code was
causing the test failure. The hash function in the snappy compression algorithm
was causing a hash table collision during compression on big-endian systems.
Compression produced different output than the equivalent code in GO causing the
test case failure. The solution implemented was to byte swap the input into the hash
function on big-endian systems as to match the little-endian input.

This test is more of a sanity check, however, it's important since it ensures that the
equivalent GO code is still producing the correct output (even on big-endian systems).
This fix should not affect the test on little-endian systems, this was confirmed
through testing on intel x64.